@@ -1,6 +1,9 @@
-"use strict";
-
-import React, { Component, PropTypes } from "react";
+// @flow
+/**
+ * Created by Konstantin Yakushin.
+ * react-native-underline-tabbar
+ */
+import React, { Component } from 'react';
 import {
   StyleSheet,
   Text,
@@ -8,87 +11,52 @@ import {
   View,
   Animated,
   ScrollView,
-  Dimensions
-} from "react-native";
+  Dimensions,
+} from 'react-native';
 
-const SCREEN_WIDTH = Dimensions.get("window").width;
+import styles from './Styles';
 
-const styles = StyleSheet.create({
-  tab: {
-    flexDirection: "row",
-    justifyContent: "center",
-    marginLeft: 20,
-    paddingBottom: 10
-  },
-  scrollContainer: {
-    paddingRight: 20
-  },
-  tabs: {
-    marginTop: 15,
-    flexDirection: "row",
-    justifyContent: "flex-start",
-    alignItems: "flex-start",
-    borderWidth: 1,
-    borderTopWidth: 0,
-    borderLeftWidth: 0,
-    borderRightWidth: 0,
-    borderBottomColor: "#d2d2d2"
-  },
+const SCREEN_WIDTH = Dimensions.get('window').width;
 
-  badgeBubble: {
-    marginTop: 4,
-    marginLeft: 5,
-    height: 12,
-    width: 17,
-    borderRadius: 4.5,
-    alignItems: "center",
-    justifyContent: "center"
-  },
-  badgeText: {
-    color: "white",
-    fontSize: 9,
-    fontWeight: "bold",
-    backgroundColor: "transparent",
-    top: -0.5
-  }
-});
+
+type TabBarProps = {
+  goToPage: Function,
+  activeTab: number,
+  tabs: Array<{ label: string, badge: string, badgeColor?: string }>,
+  underlineColor: string,
+  backgroundColor: string,
+  activeTextColor: string,
+  inactiveTextColor: string,
+  scrollContainerStyle?: StyleSheet.Styles,
+  tabStyles?: StyleSheet.Styles,
+};
+
+type TabBarState = {
+  renderUnderline: boolean,
+  tabScrollValue: number,
+};
 
 class TabBar extends Component {
-  static propTypes = {
-    goToPage: PropTypes.func,
-    activeTab: PropTypes.number,
-    tabs: PropTypes.array,
-    underlineColor: PropTypes.string,
-    backgroundColor: PropTypes.string,
-    activeTextColor: PropTypes.string,
-    inactiveTextColor: PropTypes.string,
-    scrollContainerStyle: PropTypes.object,
-    tabStyles: PropTypes.object
-  };
-
+  props: TabBarProps;
   static defaultProps = {
-    tabStyles: {}
+    tabStyles: {},
+    scrollContainerStyle: {},
   };
-
-  constructor(props) {
-    super(props);
-    this.tabState = {};
-    this.state = {
-      renderUnderline: false,
-      tabScrollValue: 0
-    };
-  }
-
-  componentDidUpdate(prevProps, prevState) {
+  tabState: Object = {};
+  scrollTabs: ?Element = null;
+  state: TabBarState = {
+    renderUnderline: false,
+    tabScrollValue: 0,
+  };
+  componentDidUpdate(prevProps: TabBarProps) {
     if (prevProps.activeTab !== this.props.activeTab) {
-      this._checkViewportOverflows();
+      this.checkViewportOverflows();
     }
   }
 
-  _checkViewportOverflows() {
-    const getScreenMargin = props => {
-      return StyleSheet.flatten([styles.tab, props.tabStyles.tab]).marginLeft;
-    };
+  checkViewportOverflows() {
+    const getScreenMargin = props =>
+      StyleSheet.flatten([styles.tab, props.tabStyles.tab]).marginLeft;
 
     const screenMargin = getScreenMargin(this.props);
     const currentTabLayout = this.tabState[this.props.activeTab];
@@ -113,10 +81,11 @@ class TabBar extends Component {
       const y = 0;
       return this.scrollTabs.scrollTo({ x, y });
     }
+    return null;
   }
 
   onTabLayout(event, page) {
-    var { x, y, width, height } = event.nativeEvent.layout;
+    const { x, y, width, height } = event.nativeEvent.layout;
     this.tabState[page] = { x, y, width, height };
     if (this.props.tabs.length === Object.keys(this.tabState).length) {
       this.setState({ renderUnderline: true });
@@ -127,8 +96,8 @@ class TabBar extends Component {
     const { activeTab, tabBadgeColor } = this.props;
     const { label, badge, badgeColor } = tab;
     const isTabActive = activeTab === page;
-    const activeTextColor = this.props.activeTextColor || "navy";
-    const inactiveTextColor = this.props.inactiveTextColor || "black";
+    const activeTextColor = this.props.activeTextColor || 'navy';
+    const inactiveTextColor = this.props.inactiveTextColor || 'black';
     const textStyle = this.props.tabBarTextStyle || {};
     return (
       <TouchableOpacity
@@ -141,9 +110,9 @@ class TabBar extends Component {
           style={[
             {
               color: isTabActive ? activeTextColor : inactiveTextColor,
-              fontWeight: isTabActive ? "400" : "400"
+              fontWeight: isTabActive ? '400' : '400',
             },
-            textStyle
+            textStyle,
           ]}
         >
           {label}
@@ -154,7 +123,7 @@ class TabBar extends Component {
             style={[
               styles.badgeBubble,
               this.props.tabStyles.badgeBubble,
-              { backgroundColor: badgeColor || activeTextColor }
+              { backgroundColor: badgeColor || activeTextColor },
             ]}
           >
             <Text style={[styles.badgeText, this.props.tabStyles.badgeText]}>
@@ -166,32 +135,32 @@ class TabBar extends Component {
   };
 
   renderUnderline() {
-    var inputRange = Object.keys(this.tabState);
-    var outputRangeLeft = [];
-    var outputRangeWidth = [];
+    const inputRange = Object.keys(this.tabState);
+    const outputRangeLeft = [];
+    const outputRangeWidth = [];
 
-    for (var k in this.tabState) {
+    for (const k in this.tabState) {
       if (this.tabState.hasOwnProperty(k)) {
         outputRangeLeft.push(this.tabState[k].x);
         outputRangeWidth.push(this.tabState[k].width);
       }
     }
 
-    var left = this.props.scrollValue.interpolate({
-      inputRange: inputRange,
-      outputRange: outputRangeLeft
+    const left = this.props.scrollValue.interpolate({
+      inputRange,
+      outputRange: outputRangeLeft,
     });
 
-    var width = this.props.scrollValue.interpolate({
-      inputRange: inputRange,
-      outputRange: outputRangeWidth
+    const width = this.props.scrollValue.interpolate({
+      inputRange,
+      outputRange: outputRangeWidth,
     });
 
-    var tabUnderlineStyle = {
-      position: "absolute",
-      backgroundColor: this.props.underlineColor || "navy",
+    const tabUnderlineStyle = {
+      position: 'absolute',
+      backgroundColor: this.props.underlineColor || 'navy',
       height: 2,
-      bottom: 0
+      bottom: 0,
     };
 
     return <Animated.View style={[tabUnderlineStyle, { left }, { width }]} />;
@@ -204,14 +173,14 @@ class TabBar extends Component {
           styles.tabs,
           { backgroundColor: this.props.backgroundColor },
           this.props.style,
-          this.props.tabBarStyle
+          this.props.tabBarStyle,
         ]}
       >
         <ScrollView
-          horizontal={true}
+          horizontal
           contentContainerStyle={[
             styles.scrollContainer,
-            this.props.scrollContainerStyle
+            this.props.scrollContainerStyle,
           ]}
           showsHorizontalScrollIndicator={false}
           ref={node => this.scrollTabs = node}
@@ -228,4 +197,4 @@ class TabBar extends Component {
   }
 }
 
-module.exports = TabBar;
+export default TabBar;
